Add dev counters
Adds hook handlers for PageSaveComplete and RollbackComplete, so that
edits and undo and rollback actions call the appropriate callbacks
(onEditSuccess or onRevert).

Adds two counters for development and testing, one that tracks all edits,
and one that tracks all app edits.

Adds a maintenance script to register the configured counters in the
counter keys table.

Change-Id: If559e5170f5cc032aace728fac8b457c7456aae3